+++ src/stream_ml/core/_multi/independent.py
@@ -16,7 +16,7 @@ from stream_ml.core.utils.cached_property import cached_property
 from stream_ml.core.utils.funcs import get_prefixed_kwargs
 
 if TYPE_CHECKING:
-    from stream_ml.core.data import Data
+    from stream_ml.core._data import Data
     from stream_ml.core.params._core import ModelParameter
     from stream_ml.core.typing import ParamNameAllOpts, ParamsLikeDict
     from stream_ml.core.utils.frozen_dict import FrozenDict
