+++ src/stream_ml/core/params/bounds/_base.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
     from collections.abc import Iterator
 
     from stream_ml.core._core.api import Model
-    from stream_ml.core.data import Data
+    from stream_ml.core._data import Data
     from stream_ml.core.params import Params
     from stream_ml.core.typing import NNModel
