+++ src/stream_ml/core/params/bounds/_builtin.py
@@ -14,7 +14,7 @@ from stream_ml.core.utils.compat import array_at, copy
 
 if TYPE_CHECKING:
     from stream_ml.core._core.api import Model
-    from stream_ml.core.data import Data
+    from stream_ml.core._data import Data
     from stream_ml.core.params._values import Params
     from stream_ml.core.params.scaler import ParamScaler
     from stream_ml.core.typing import NNModel
