+++ src/stream_ml/core/prior/_core.py
@@ -12,7 +12,7 @@ from stream_ml.core.typing import Array
 
 if TYPE_CHECKING:
     from stream_ml.core._core.api import Model
-    from stream_ml.core.data import Data
+    from stream_ml.core._data import Data
     from stream_ml.core.params._values import Params
     from stream_ml.core.typing import NNModel
