+++ src/stream_ml/core/prior/_weight.py
@@ -16,8 +16,8 @@ from stream_ml.core.utils.scale import DataScaler  # noqa: TCH001
 
 if TYPE_CHECKING:
     from stream_ml.core._core.api import Model
+    from stream_ml.core._data import Data
     from stream_ml.core._multi.bases import ModelsBase
-    from stream_ml.core.data import Data
     from stream_ml.core.params._values import Params
     from stream_ml.core.typing import NNModel
